Version 1.0.0
As far as I can tell, I've removed most, if not all of the bugs from my
application. There may be an issue with custom commands which I haven't
discovered yet, but I believe everything else is correct. For the time
being, this build is feature-complete, and I don't intend to add any new
features. Any subsequent builds should be bug fixes only.